Put the PythonTranslator object into its own file
